First commit of the cyffld2 project
I'm just committing this as I am about to leave from work. So
far this just implements a very basic fitting process. I haven't
linked the submodule of ffld2 yet because the projects are tightly
coupled and I want to make the changes in real-time (so I'm just
symlinking it at the moment).