Windows build attempt
Attempting to get this building on Windows, but it has a load of nasty
dependencies that don't play nicely on Windows. This includes libxml2
and fftw3. Unfortunately, at the moment, fftw3 is built with mingw and
seems to cause a segfault at runtime. Building works, but running
doesn't.

These changes are highly tied to the other recipes we maintain for
libxml2 and fftw3 on Windows.