Another fix for Windows...
Going around in circles a bit, but this seems to finally work.
The problem was the mixing of MSVC and mingw C-libraries at
run time - see this: http://www.fftw.org/install/windows.html
for more information (the DLLs and wisdom bit). Also, look
at the comment in Patchwork.h inside the FFLD submodule.